requirements.dev.txt
pdbpp==0.10.2
pytest==5.2.0
prospector==1.1.5
pylint==2.1.1
ipython==7.10.1
pytest-django==3.7.0
